Refactor MemoryUsageDetails to be a view model
